+++ driven/storage/adapter.go
@@ -118,6 +118,7 @@ func (sa *Adapter) RegisterStorageListener(storageListener Listener) {
 
 // PerformTransaction performs a transaction
 func (sa *Adapter) PerformTransaction(transaction func(context TransactionContext) error) error {
+	//TODO: may want to add timeout arg to PerformTransaction to be able to specify individual timeouts (mainly for slow operations)
 	// transaction
 	err := sa.db.dbClient.UseSession(context.Background(), func(sessionContext mongo.SessionContext) error {
 		err := sessionContext.StartTransaction()
@@ -2229,17 +2230,10 @@ func (sa *Adapter) InsertPermissions(context TransactionContext, items []model.P
 		stgPermissions[i] = p
 	}
 
-	var res *mongo.InsertManyResult
-	var err error
-	if context != nil {
-		res, err = sa.db.permissions.InsertManyWithContext(context, stgPermissions, nil)
-	} else {
-		res, err = sa.db.permissions.InsertMany(stgPermissions, nil)
-	}
+	res, err := sa.db.permissions.InsertManyWithContext(context, stgPermissions, nil)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionInsert, model.TypePermission, nil, err)
 	}
-
 	if len(res.InsertedIDs) != len(items) {
 		return errors.ErrorAction(logutils.ActionInsert, model.TypePermission, &logutils.FieldArgs{"inserted": len(res.InsertedIDs), "expected": len(items)})
 	}
@@ -2252,28 +2246,82 @@ func (sa *Adapter) UpdatePermission(context TransactionContext, item model.Permi
 	//TODO
 	//This will be slow operation as we keep a copy of the entity in the users collection without index.
 	//Maybe we need to up the transaction timeout for this operation because of this.
-	//TODO
-	//Update the permission in all collection where there is a copy of it - accounts, application_roles, application_groups, service_accounts
+	if context == nil {
+		transaction := func(newContext TransactionContext) error {
+			return sa.updatePermission(newContext, item)
+		}
+		return sa.PerformTransaction(transaction)
+	}
 
-	// Update serviceIDs
-	filter := bson.D{primitive.E{Key: "name", Value: item.Name}}
+	return sa.updatePermission(context, item)
+}
 
-	now := time.Now().UTC()
+func (sa *Adapter) updatePermission(context TransactionContext, item model.Permission) error {
+	//update permission
+	permissionFilter := bson.D{primitive.E{Key: "name", Value: item.Name}}
 	permissionUpdate := bson.D{
 		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "description", Value: item.Description},
 			primitive.E{Key: "service_id", Value: item.ServiceID},
 			primitive.E{Key: "assigners", Value: item.Assigners},
-			primitive.E{Key: "date_updated", Value: &now},
+			primitive.E{Key: "date_updated", Value: item.DateUpdated},
 		}},
 	}
 
-	res, err := sa.db.permissions.UpdateOneWithContext(context, filter, permissionUpdate, nil)
+	res, err := sa.db.permissions.UpdateOneWithContext(context, permissionFilter, permissionUpdate, nil)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypePermission, &logutils.FieldArgs{"name": item.Name}, err)
 	}
-
 	if res.ModifiedCount != 1 {
-		return errors.ErrorAction(logutils.ActionUpdate, model.TypePermission, logutils.StringArgs("unexpected modified count"))
+		return errors.ErrorAction(logutils.ActionUpdate, model.TypePermission, &logutils.FieldArgs{"name": item.Name, "modified": res.ModifiedCount, "expected": 1})
+	}
+
+	// update all roles that have the permission
+	key := "permissions.name"
+	roles, err := sa.findAppOrgRoles(context, &key, item.Name, "")
+	for _, r := range roles {
+		for pidx, p := range r.Permissions {
+			if p.Name == item.Name {
+				r.Permissions[pidx] = item
+				err = sa.UpdateAppOrgRole(context, r)
+				break
+			}
+		}
+	}
+
+	// update all groups that have the permission
+	groups, err := sa.findAppOrgGroups(context, &key, item.Name, "")
+	for _, g := range groups {
+		for pidx, p := range g.Permissions {
+			if p.Name == item.Name {
+				g.Permissions[pidx] = item
+				err = sa.UpdateAppOrgGroup(context, g)
+				break
+			}
+		}
+	}
+
+	//update all roles, groups, accounts, and service accounts that have the permission
+	dependentsFilter := bson.D{primitive.E{Key: "permissions.name", Value: item.Name}}
+	dependentsUpdate := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "permissions.$.description", Value: item.Description},
+			primitive.E{Key: "permissions.$.service_id", Value: item.ServiceID},
+			primitive.E{Key: "permissions.$.assigners", Value: item.Assigners},
+			primitive.E{Key: "permissions.$.date_updated", Value: item.DateUpdated},
+		}},
+	}
+
+	//accounts
+	res, err = sa.db.accounts.UpdateManyWithContext(context, dependentsFilter, dependentsUpdate, nil)
+	if err = sa.getUpdateManyError(res, err, model.TypeAccount, "permissions.name", item.Name); err != nil {
+		return err
+	}
+
+	//service accounts
+	res, err = sa.db.serviceAccounts.UpdateManyWithContext(context, dependentsFilter, dependentsUpdate, nil)
+	if err = sa.getUpdateManyError(res, err, model.TypeServiceAccount, "permissions.name", item.Name); err != nil {
+		return err
 	}
 
 	return nil
@@ -2289,22 +2337,7 @@ func (sa *Adapter) DeletePermission(id string) error {
 
 // FindAppOrgRoles finds all application organization roles fora given AppOrg ID
 func (sa *Adapter) FindAppOrgRoles(appOrgID string) ([]model.AppOrgRole, error) {
-	rolesFilter := bson.D{primitive.E{Key: "app_org_id", Value: appOrgID}}
-	var rolesResult []appOrgRole
-	err := sa.db.applicationsOrganizationsRoles.Find(rolesFilter, &rolesResult, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	//get the application organization from the cached ones
-	appOrg, err := sa.getCachedApplicationOrganizationByKey(appOrgID)
-	if err != nil {
-		return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeOrganization, &logutils.FieldArgs{"app_org_id": appOrg}, err)
-	}
-
-	result := appOrgRolesFromStorage(rolesResult, *appOrg)
-
-	return result, nil
+	return sa.findAppOrgRoles(nil, nil, "", appOrgID)
 }
 
 // FindAppOrgRolesByIDs finds a set of application organization roles for the provided IDs
@@ -2357,18 +2390,42 @@ func (sa *Adapter) FindAppOrgRole(id string, appOrgID string) (*model.AppOrgRole
 	return &result, nil
 }
 
-// InsertAppOrgRole inserts a new application organization role
-func (sa *Adapter) InsertAppOrgRole(context TransactionContext, item model.AppOrgRole) error {
-	appOrg, err := sa.getCachedApplicationOrganizationByKey(item.AppOrg.ID)
-	if err != nil {
-		return errors.WrapErrorData(logutils.StatusMissing, model.TypeApplicationOrganization, &logutils.FieldArgs{"app_org_id": item.AppOrg.ID}, err)
+func (sa *Adapter) findAppOrgRoles(context TransactionContext, key *string, id string, appOrgID string) ([]model.AppOrgRole, error) {
+	filter := bson.D{}
+	errFields := logutils.FieldArgs{}
+	if key != nil {
+		filter = append(filter, primitive.E{Key: *key, Value: id})
+		errFields[*key] = id
 	}
-	if appOrg == nil {
-		return errors.ErrorData(logutils.StatusMissing, model.TypeApplicationOrganization, &logutils.FieldArgs{"app_org_id": item.AppOrg.ID})
+	if appOrgID != "" {
+		filter = append(filter, primitive.E{Key: "app_org_id", Value: appOrgID})
+		errFields["app_org_id"] = appOrgID
 	}
 
+	var rolesResult []appOrgRole
+	err := sa.db.applicationsOrganizationsRoles.FindWithContext(context, filter, &rolesResult, nil)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAppOrgRole, &errFields, err)
+	}
+
+	var result []model.AppOrgRole
+	if len(rolesResult) > 0 {
+		appOrg, err := sa.getCachedApplicationOrganizationByKey(rolesResult[0].AppOrgID)
+		if err != nil || appOrg == nil {
+			return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeApplicationOrganization, &logutils.FieldArgs{"app_org_id": rolesResult[0].AppOrgID}, err)
+		}
+		result = appOrgRolesFromStorage(rolesResult, *appOrg)
+	} else {
+		result = make([]model.AppOrgRole, 0)
+	}
+
+	return result, nil
+}
+
+// InsertAppOrgRole inserts a new application organization role
+func (sa *Adapter) InsertAppOrgRole(context TransactionContext, item model.AppOrgRole) error {
 	role := appOrgRoleToStorage(item)
-	_, err = sa.db.applicationsOrganizationsRoles.InsertOneWithContext(context, role)
+	_, err := sa.db.applicationsOrganizationsRoles.InsertOneWithContext(context, role)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionInsert, model.TypeAppOrgRole, nil, err)
 	}
@@ -2376,11 +2433,72 @@ func (sa *Adapter) InsertAppOrgRole(context TransactionContext, item model.AppOr
 }
 
 // UpdateAppOrgRole updates application organization role
-func (sa *Adapter) UpdateAppOrgRole(item model.AppOrgRole) error {
-	//TODO
-	//This will be slow operation as we keep a copy of the entity in the users collection without index.
-	//Maybe we need to up the transaction timeout for this operation because of this.
-	return errors.New(logutils.Unimplemented)
+func (sa *Adapter) UpdateAppOrgRole(context TransactionContext, item model.AppOrgRole) error {
+	if context == nil {
+		transaction := func(newContext TransactionContext) error {
+			return sa.updateAppOrgRole(newContext, item)
+		}
+		return sa.PerformTransaction(transaction)
+	}
+
+	return sa.updateAppOrgRole(context, item)
+}
+
+func (sa *Adapter) updateAppOrgRole(context TransactionContext, item model.AppOrgRole) error {
+	// update role
+	roleFilter := bson.D{primitive.E{Key: "_id", Value: item.ID}}
+	roleUpdate := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "name", Value: item.Name},
+			primitive.E{Key: "description", Value: item.Description},
+			primitive.E{Key: "permissions", Value: item.Permissions},
+			primitive.E{Key: "system", Value: item.System},
+			primitive.E{Key: "date_updated", Value: item.DateUpdated},
+		}},
+	}
+
+	res, err := sa.db.applicationsOrganizationsRoles.UpdateOneWithContext(context, roleFilter, roleUpdate, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAppOrgRole, &logutils.FieldArgs{"id": item.ID}, err)
+	}
+	if res.ModifiedCount != 1 {
+		return errors.ErrorAction(logutils.ActionUpdate, model.TypeAppOrgRole, &logutils.FieldArgs{"id": item.ID, "modified": res.ModifiedCount, "expected": 1})
+	}
+
+	// update all groups that have the role
+	key := "roles._id"
+	groups, err := sa.findAppOrgGroups(context, &key, item.ID, item.AppOrg.ID)
+	for _, g := range groups {
+		for ridx, r := range g.Roles {
+			if r.ID == item.ID {
+				g.Roles[ridx] = item
+				err = sa.UpdateAppOrgGroup(context, g)
+				break
+			}
+		}
+	}
+
+	// update all accounts that have the role
+	accountsFilter := bson.D{primitive.E{Key: "roles.role._id", Value: item.ID}}
+	accountsUpdate := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "roles.$.role.name", Value: item.Name},
+			primitive.E{Key: "roles.$.role.description", Value: item.Description},
+			primitive.E{Key: "roles.$.role.permissions", Value: item.Permissions},
+			primitive.E{Key: "roles.$.role.system", Value: item.System},
+			primitive.E{Key: "roles.$.role.date_updated", Value: item.DateUpdated},
+		}},
+	}
+
+	res, err = sa.db.accounts.UpdateManyWithContext(context, accountsFilter, accountsUpdate, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccount, &logutils.FieldArgs{"roles.role._id": item.ID}, err)
+	}
+	if res.ModifiedCount != res.MatchedCount {
+		return errors.ErrorAction(logutils.ActionUpdate, model.TypeAccount, &logutils.FieldArgs{"roles.role._id": item.ID, "modified": res.ModifiedCount, "expected": res.MatchedCount})
+	}
+
+	return nil
 }
 
 // DeleteAppOrgRole deletes application organization role
@@ -2425,21 +2543,7 @@ func (sa *Adapter) InsertAppOrgRolePermissions(context TransactionContext, roleI
 
 // FindAppOrgGroups finds all application organization groups for the provided AppOrg ID
 func (sa *Adapter) FindAppOrgGroups(appOrgID string) ([]model.AppOrgGroup, error) {
-	filter := bson.D{primitive.E{Key: "app_org_id", Value: appOrgID}}
-	var groupsResult []appOrgGroup
-	err := sa.db.applicationsOrganizationsGroups.Find(filter, &groupsResult, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	appOrg, err := sa.getCachedApplicationOrganizationByKey(appOrgID)
-	if err != nil {
-		return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeOrganization, &logutils.FieldArgs{"app_org_id": appOrg}, err)
-	}
-
-	result := appOrgGroupsFromStorage(groupsResult, *appOrg)
-
-	return result, nil
+	return sa.findAppOrgGroups(nil, nil, "", appOrgID)
 }
 
 // FindAppOrgGroupsByIDs finds a set of application organization groups for the provided IDs
@@ -2491,6 +2595,38 @@ func (sa *Adapter) FindAppOrgGroup(id string, appOrgID string) (*model.AppOrgGro
 	return &result, nil
 }
 
+func (sa *Adapter) findAppOrgGroups(context TransactionContext, key *string, id string, appOrgID string) ([]model.AppOrgGroup, error) {
+	filter := bson.D{}
+	errFields := logutils.FieldArgs{}
+	if key != nil {
+		filter = append(filter, primitive.E{Key: *key, Value: id})
+		errFields[*key] = id
+	}
+	if appOrgID != "" {
+		filter = append(filter, primitive.E{Key: "app_org_id", Value: appOrgID})
+		errFields["app_org_id"] = appOrgID
+	}
+
+	var groupsResult []appOrgGroup
+	err := sa.db.applicationsOrganizationsGroups.FindWithContext(context, filter, &groupsResult, nil)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAppOrgGroup, &errFields, err)
+	}
+
+	var result []model.AppOrgGroup
+	if len(groupsResult) > 0 {
+		appOrg, err := sa.getCachedApplicationOrganizationByKey(groupsResult[0].AppOrgID)
+		if err != nil || appOrg == nil {
+			return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeApplicationOrganization, &logutils.FieldArgs{"app_org_id": groupsResult[0].AppOrgID}, err)
+		}
+		result = appOrgGroupsFromStorage(groupsResult, *appOrg)
+	} else {
+		result = make([]model.AppOrgGroup, 0)
+	}
+
+	return result, nil
+}
+
 // InsertAppOrgGroup inserts a new application organization group
 func (sa *Adapter) InsertAppOrgGroup(context TransactionContext, item model.AppOrgGroup) error {
 	group := appOrgGroupToStorage(item)
@@ -2503,11 +2639,63 @@ func (sa *Adapter) InsertAppOrgGroup(context TransactionContext, item model.AppO
 }
 
 // UpdateAppOrgGroup updates application organization group
-func (sa *Adapter) UpdateAppOrgGroup(item model.AppOrgGroup) error {
-	//TODO
-	//This will be slow operation as we keep a copy of the entity in the users collection without index.
-	//Maybe we need to up the transaction timeout for this operation because of this.
-	return errors.New(logutils.Unimplemented)
+func (sa *Adapter) UpdateAppOrgGroup(context TransactionContext, item model.AppOrgGroup) error {
+	if context == nil {
+		transaction := func(newContext TransactionContext) error {
+			return sa.updateAppOrgGroup(newContext, item)
+		}
+		return sa.PerformTransaction(transaction)
+	}
+
+	return sa.updateAppOrgGroup(context, item)
+}
+
+func (sa *Adapter) updateAppOrgGroup(context TransactionContext, item model.AppOrgGroup) error {
+	roles := appOrgRolesToStorage(item.Roles)
+
+	// update group
+	groupFilter := bson.D{primitive.E{Key: "_id", Value: item.ID}}
+	groupUpdate := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "name", Value: item.Name},
+			primitive.E{Key: "description", Value: item.Description},
+			primitive.E{Key: "permissions", Value: item.Permissions},
+			primitive.E{Key: "roles", Value: roles},
+			primitive.E{Key: "system", Value: item.System},
+			primitive.E{Key: "date_updated", Value: item.DateUpdated},
+		}},
+	}
+
+	res, err := sa.db.applicationsOrganizationsGroups.UpdateOneWithContext(context, groupFilter, groupUpdate, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAppOrgGroup, &logutils.FieldArgs{"id": item.ID}, err)
+	}
+	if res.ModifiedCount != 1 {
+		return errors.ErrorAction(logutils.ActionUpdate, model.TypeAppOrgGroup, &logutils.FieldArgs{"id": item.ID, "modified": res.ModifiedCount, "expected": 1})
+	}
+
+	// update all accounts that have the group
+	accountsFilter := bson.D{primitive.E{Key: "groups.group._id", Value: item.ID}}
+	accountsUpdate := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "groups.$.group.name", Value: item.Name},
+			primitive.E{Key: "groups.$.group.description", Value: item.Description},
+			primitive.E{Key: "groups.$.group.permissions", Value: item.Permissions},
+			primitive.E{Key: "groups.$.group.roles", Value: roles},
+			primitive.E{Key: "groups.$.group.system", Value: item.System},
+			primitive.E{Key: "groups.$.group.date_updated", Value: item.DateUpdated},
+		}},
+	}
+
+	res, err = sa.db.accounts.UpdateManyWithContext(context, accountsFilter, accountsUpdate, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccount, &logutils.FieldArgs{"groups.group.id": item.ID}, err)
+	}
+	if res.ModifiedCount != res.MatchedCount {
+		return errors.ErrorAction(logutils.ActionUpdate, model.TypeAccount, &logutils.FieldArgs{"groups.group.id": item.ID, "modified": res.ModifiedCount, "expected": res.MatchedCount})
+	}
+
+	return nil
 }
 
 // DeleteAppOrgGroup deletes application organization group
@@ -3289,6 +3477,16 @@ func (sa *Adapter) abortTransaction(sessionContext mongo.SessionContext) {
 	if err != nil {
 		sa.logger.Errorf("error aborting a transaction - %s", err)
 	}
+}
+
+func (sa *Adapter) getUpdateManyError(res *mongo.UpdateResult, err error, dataType logutils.MessageDataType, key string, value string) error {
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, dataType, &logutils.FieldArgs{key: value}, err)
+	}
+	if res.ModifiedCount != res.MatchedCount {
+		return errors.ErrorAction(logutils.ActionUpdate, dataType, &logutils.FieldArgs{key: value, "modified": res.ModifiedCount, "expected": res.MatchedCount})
+	}
+	return nil
 }
 
 // NewStorageAdapter creates a new storage adapter instance