+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Update role permissions admin API [#516](https://github.com/rokwire/core-building-block/issues/516)
+- Update group admin API [#518](https://github.com/rokwire/core-building-block/issues/518)
 - Allow passing nil context to WithContext storage functions [#494](https://github.com/rokwire/core-building-block/issues/494)
+
 ### Fixed
+- Update permission should change roles/groups/accounts [#542](https://github.com/rokwire/core-building-block/issues/542)
 - Authorization policy comments not working [#506](https://github.com/rokwire/core-building-block/issues/506)
 
 ## [1.24.1] - 2022-07-07